Add detection for Switch and China UnionPay networks
Add support for Switch, China UnionPay to detectNetwork function.

To resolve conflicts with Visa, must check for Switch numbers
first in our control flow.

Because prefix ranges have greatly increased with China UnionPay,
it's no longer efficient to use Array and indexOf methods.

Extend Number prototype with a helper function to find if a number
is between a range, inclusive of the lower and upper bounds.